+++ apps/coordinator-api/src/app/contexts/wallet/services/wallet_service.py
@@ -7,6 +7,7 @@ Service for managing agent wallets across multiple blockchain networks.
 from __future__ import annotations
 
 import secrets
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import select
@@ -80,7 +81,7 @@ class WalletService:
         """Get all tracked balances for a wallet"""
         return self.session.execute(select(TokenBalance).where(TokenBalance.wallet_id == wallet_id)).all()  # type: ignore[arg-type, return-value]
 
-    async def update_balance(self, wallet_id: int, chain_id: int, token_address: str, balance: float) -> TokenBalance:
+    async def update_balance(self, wallet_id: int, chain_id: int, token_address: str, balance: Decimal) -> TokenBalance:
         """Update a specific token balance for a wallet"""
         record = self.session.execute(
             select(TokenBalance).where(
